Create table code
This is a rundown of the table to go through one char at a time and determine the token by using states. It needs work and needs a while loop and a way to keep track of the chars to output when token is determined once it hits a final state.